Added Automatic movement option to unity

+++ UnitySimulation/Assets/Scripts/PythonModels/FaceCount.cs
@@ -3,5 +3,5 @@ using System;
 [Serializable]
 public class FaceCount
 {
-    public float count;
+    public int count;
 }
